Moved variables fixture to a separate file
